+++ CHANGELOG.md
@@ -13,11 +13,13 @@ The minor version will be incremented upon a breaking change and the patch versi
 ### Features
 
 ### Fixes
+
 - client: Fix deadlock when having multiple websocket listeners ([#4250](https://github.com/solana-foundation/anchor/pull/4250)).
 
 ### Breaking
 
 - lang: Rename `errors` and `ProgramError` of `declare_program!` ([#4347](https://github.com/solana-foundation/anchor/pull/4347)).
+- client: Remove the `solana-account-decoder` crate export ([#4373](https://github.com/solana-foundation/anchor/pull/4373)).
 
 ## [1.0.0-rc.5] - 2026-03-20